requirements.txt
google-generativeai
python-dotenv
transformers
nltk
streamlit
pytest
torch